Bump mocha upper bound
The version of mocha we were pinning to (1.20.1), allowed any
version of the debug dependency to be installed. This meant that
the latest version was being pulled in. However the latest version
was not compatible with Node 4 (see https://github.com/visionmedia/debug/issues/603).

This change bumps mocha to a version (1.21.5) where they constrain debug to
a version (2.0.0, that is Node 4 compatible)